Modify Torch Support to handle tensors on GPU
It seems the torch support assumed tensors exist on CPU, as this is a pre-requisite for conversion to numpy arrays with `some_tensor.numpy()`. Perhaps this was an intentional design choice to avoid occupying non-accelerator memory for users who aren't being careful, but moving non-cpu tensors onto cpu automatically for the purposes of visualization is likely acceptable in most use cases, so I have added this.