Simplify torch handling to support more backends.
Disables named-axis renaming if the tensor is already nameless, and
replaces calls to the `isposinf`/`isneginf` primitives with simple
equality checks. This reduces the set of operations that tensors
must support, allowing Treescope to support more backends.

With this change it should be possible to use Treescope with torch
tensors that use the "mps" (Metal Performance Shaders / Apple Silicon)
device backend.

PiperOrigin-RevId: 668141265