Show requires_grad info for torch tensors and parameters.
For `Tensor`s, we add requires_grad=True to the summary when
that flag is set on the tensor.

For `nn.Parameter`s, we assume gradients are required by default
and add requires_grad=False when the parameter is frozen.

Summary shows regardless of whether or not autovizualization is on.